State walkthrough goals upfront
This tweaks the introduction a little, and adds a paragraph to say
that the end goal is a cloneable dataset on Nextcloud, and emphasises
the consumer benefits. The post is taken out of draft mode.

Admittedly, the walkthrough is lengthier then absolutely
necessary (all config could be applied at once, and access tested just
once), but it is my natural approach of "let's see what these things
do in isolation" shining through. I hope this is OK, and that with the
new introduction the goal does not seem too vague.

Suggested-by: Michael Hanke <michael.hanke@gmail.com>