Update PaperMod theme
This updates the PaperMod theme from v8.0~5 (5 commits before 8.0) to
26 commits past v8.0 (i.e. current master branch).

The main appeal of the update is Papermod's recent inclusion of the
opengraph 'fediverse:creator' meta tag. Mastodon recognizes this tag
in linked pages and displays a link to creator's Mastodon profile if
the given domain is added to the profile settings. The meta tag will
be rendered in posts if they have the following in the front matter:

```
social:
  fediverse_creator: "@username@mastodon.social"
```

This can be useful for us considering the multi-author nature of the
DataLad blog and the fact that our project uses Mastodon as its main
social media channel. Setting the parameter would be totally optional
for authors. In addition to setting their handle in metadata, the
author would need to add blog.datalad.org to the list of pages which
can attribute ownership to them (profile settings).

This is a relatively new Mastodon feature, introduced in summer
2024. More in Mastodon's blog:
https://blog.joinmastodon.org/2024/07/highlighting-journalism-on-mastodon/

If the PaperMod's README is to be trusted, this bumps the minimal
required hugo version from 0.112.4 to 0.125.7 which seems
reasonable (e.g. Fedora currently ships 0.126.2, IIRC our Forgejo
actions currently use 0.131, and the bleeding edge is 0.145).