Add cover and tags
This is not the most creative cover, but I did not have the skill and
inspiration to draw a Git repository in an annex remote on
Nextcloud. The (carefully staged) screenshot will have to do.

I assembled the image in GIMP, so I am including the xcf source
instead of the customary svg, hope it's OK. Size-wise it is in line
with all the others.

I was a little bit hesitant about including the nextcloud logo
considering how detailed their guidelines are, but I believe that the
use constitutes "fair use" and also falls under "Publications" (and,
to a lesser degree, "Advocacy groups") in their trademark
guidelines. The logo is taken from their own press materials:

- https://nextcloud.com/press/
- https://nextcloud.com/brand/
- https://nextcloud.com/trademarks/